add money package
Added an integer based package for handling USD based on cents rather
than floating point dollars.

+++ main.go
@@ -10,7 +10,7 @@ import (
 
 type Loan struct {
 	name        string
-	balance     int
+	balance     float64
 	rate        float64
 	min_payment float64
 	due_day     int
